+++ humilis_push_processor/__init__.py
@@ -1,3 +1,3 @@
 """A Lambda processor for push event notifications."""
 
-__version__ = "0.0.3"
+__version__ = "0.0.6"