+++ tests/unit/test_lambda.py
@@ -1,9 +1,13 @@
 """Test the logic of the Lambda function."""
+
+import copy
+
 from mock import Mock
 
 import pytest
 
 from humilis_push_processor.lambda_function.handler.processor import process_event  # noqa
+from lambdautils.exception import CriticalError
 
 
 def _identity(ev, state_args=None, **kwargs):
@@ -16,79 +20,66 @@ def _all(ev, state_args=None, **kwargs):
     return True
 
 
+def _dupper(ev, state_args=None, **kwargs):
+    """Duplicates every input event."""
+    return [ev, ev]
+
+
 def _none(ev, state_args=None, **kwargs):
     """A pass-none filter callable."""
     return False
 
 
-_input = {
-    "firehose_delivery_stream": "ifhs",
-    "mapper": Mock(side_effect=_identity),
-    "filter": Mock(side_effect=_all)}
+def _make_input(filter=None, mapper=None, kstream=None):
+    input = {}
+    if filter is not None:
+        input["filter"] = Mock(side_effect=filter)
+    if mapper is not None:
+        input["mapper"] = Mock(side_effect=mapper)
+    if kstream is not None:
+        input["kinesis_stream"] = kstream
+
+    return input
 
 
-_output = [{
-    "kinesis_stream": "oks1",
-    "firehose_delivery_stream": "ofhs1",
-    "mapper": Mock(side_effect=_identity),
-    "filter": Mock(side_effect=_all),
-    "partition_key": Mock(size_effect=lambda ev: ev.get("client_id"))
-    },
-    {
-    "kinesis_stream": "oks2",
-    "firehose_delivery_stream": "ofhs2",
-    "mapper": Mock(side_effect=_identity),
-    "filter": Mock(side_effect=_all)
-    }]
+def _make_output(filter=None, mapper=None, kstream=None, fstream=None, n=1):
+    o = {}
+    if filter is not None:
+        o["filter"] = Mock(side_effect=filter)
+    if mapper is not None:
+        o["mapper"] = Mock(side_effect=mapper)
+    if kstream is not None:
+        o["kinesis_stream"] = kstream
+    if fstream is not None:
+        o["firehose_delivery_stream"] = fstream
+
+    return [copy.deepcopy(o) for _ in range(n)]
 
 
 @pytest.mark.parametrize(
     "n,e,l,s,i,os,kput,fput", [
         [1, "e", "l", "s",
-            {
-                "kinesis_stream": "iks",
-            },
-            [
-                {
-                    "kinesis_stream": "oks1",
-                    "firehose_delivery_stream": "ofhs1",
-                    "mapper": Mock(side_effect=_identity),
-                    "filter": Mock(side_effect=_all)},
-                {
-                    "kinesis_stream": "oks2",
-                    "firehose_delivery_stream": "ofhs2",
-                    "mapper": Mock(side_effect=_identity),
-                    "filter": Mock(side_effect=_none)}
-                ],
-            1, 1],
-        [2, "e", "l", "s", _input, _output, 2, 3],
-        [10, "e", "l", "s",
-            {
-                "kinesis_stream": "iks",
-            },
-            [{
-                "kinesis_stream": "oks1",
-                "firehose_delivery_stream": "ofhs1",
-                "mapper": Mock(side_effect=_identity),
-                "filter": Mock(side_effect=_all)}],
-            1, 1],
-        [4, "e", "l", "s",
-            {
-                "kinesis_stream": "iks",
-                "filter": Mock(side_effect=_none)
-            },
-            [{
-                "kinesis_stream": "oks1",
-                "firehose_delivery_stream": "ofhs1",
-                "mapper": Mock(side_effect=_identity),
-                "filter": Mock(side_effect=_all)}],
-            0, 0],
-        [3, "e", "l", "s", _input, [], 0, 1],
+         _make_input(kstream="k"),
+         _make_output(_all, _dupper, "k", "f", 2), 2, 2],
+        [1, "e", "l", "s",
+         _make_input(kstream="k"),
+         _make_output(_all, _identity, "k", "f", 2), 2, 2],
         [5, "e", "l", "s",
-            {
-                "kinesis_stream": "iks",
-            },
-            [], 0, 0]
+         _make_input(kstream="k"),
+         _make_output(_all, _identity, "k", None, 2), 2, 0],
+        [10, "e", "l", "s",
+         _make_input(kstream="k"),
+         _make_output(_all, _identity, None, "k", 2), 0, 2],
+        [4, "e", "l", "s",
+         _make_input(kstream="k"),
+         _make_output(_all, _identity, None, None, 2), 0, 0],
+        [5, "e", "l", "s",
+         _make_input(_none, None, "k"),
+         _make_output(_all, _identity, "k", "f"), 0, 0],
+        [1, "e", "l", "s",
+         _make_input(_all, _identity, "k"),
+         _make_output(_none, _identity, "k", "f", 2), 0, 0],
+        [1, "e", "l", "s", [], [], 0, 0]
         ])
 def test_process_event(n, e, l, s, i, os, kput, fput, boto3_client,
                        monkeypatch):
@@ -99,12 +90,21 @@ def test_process_event(n, e, l, s, i, os, kput, fput, boto3_client,
     assert boto3_client("kinesis").put_records.call_count == kput
     assert boto3_client("firehose").put_record_batch.call_count == fput
 
-    ifilter = i.get("filter")
+    if i:
+        ifilter = i.get("filter")
+    else:
+        ifilter = None
+
     if ifilter:
         assert ifilter.call_count == n
         # Need to reset the call count because events is a parametrized fixture
         ifilter.reset_mock()
-    imapper = i.get("mapper")
+
+    if i:
+        imapper = i.get("mapper")
+    else:
+        imapper = None
+
     if imapper:
         if ifilter is None or ifilter.side_effect == _all:
             assert imapper.call_count == n
@@ -139,3 +139,16 @@ def test_process_event(n, e, l, s, i, os, kput, fput, boto3_client,
 
         if pk:
             pk.reset_mock()
+
+
+def test_bad_callable(events, context):
+    """Bad mappers should raise an exception."""
+
+    notification = {"Records": [{} for _ in range(len(events))]}
+
+    def bad_mapper(ev, context):
+        return "I should never return a string!"
+
+    os = [{"mapper": Mock(side_effect=bad_mapper)}]
+    with pytest.raises(CriticalError):
+        process_event(notification, context, "e", "l", "s", [], os)