+++ CHANGES.rst
@@ -1,6 +1,22 @@
 Changelog
 =========
 
+0.0.6
+-----
+
+- Support for one-to-many mappers
+
+0.0.5
+-----
+
+- Mappers do not modify input events directly
+
+0.0.4
+-----
+
+- Users can also specify input/output/error streams by name instead of as
+  references to another layer outputs.
+
 0.0.3
 -----